// packages/token-kiosk/src/auth-manager.d.ts
export type components = {
  schemas: {
    error: {
      message: string;
    };
    /** @enum {string} */
    environment: 'np' | 'stage' | 'prod';
    /**
     * Format: int32
     * @default 1
     */
    version: number;
    /** @enum {string} */
    assetType: 'TEST' | 'TEST_DATA' | 'POLICY' | 'DATA';
    name: string;
    domain: {
      name: components['schemas']['name'];
    };
    pointOfContact: {
      name: components['schemas']['name'];
      /** Format: email */
      email: string;
      /** Format: phone */
      phone: string;
    };
    namelessClient: {
      hebName: components['schemas']['name'];
      description?: string;
      branch?: string;
      /** Format: date-time */
      readonly createdAt: string;
      /** Format: date-time */
      readonly updatedAt: string;
      techPointOfContact?: components['schemas']['pointOfContact'];
      productPointOfContact?: components['schemas']['pointOfContact'];
      tags?: string[];
    };
    client: components['schemas']['namelessClient'] & {
      name: components['schemas']['name'];
    };
    connection: {
      name: components['schemas']['name'];
      environment: components['schemas']['environment'];
      version: components['schemas']['version'];
      /** @default true */
      enabled: boolean;
      token: string;
      domains: string[];
      /** Format: date-time */
      readonly createdAt?: string;
      allowNoBrowserConnection: boolean;
      allowNoOriginConnection: boolean;
      origins: string[];
    };
    publicJWK: {
      kty: string;
      n: string;
      e: string;
      alg: string;
      kid: string;
    };
    privateJWK: components['schemas']['publicJWK'] & {
      d: string;
      p: string;
      q: string;
      dp: string;
      dq: string;
      qi: string;
    };
    key: {
      publicKey: components['schemas']['publicJWK'];
      privateKey: components['schemas']['privateJWK'];
      version: components['schemas']['version'];
      environment: components['schemas']['environment'];
    };
    asset: {
      environment: components['schemas']['environment'][];
      /** Format: date-time */
      readonly createdAt: string;
      /** Format: byte */
      value: string;
      uri: string;
      type: components['schemas']['assetType'];
      /** @default false */
      isTemplate: boolean;
      version: components['schemas']['version'];
      name: string;
    };
    bundle: {
      /** Format: int32 */
      id?: number;
      hash?: string;
      metadata?: {
        [key: string]: unknown;
      };
      assets?: {
        name: string;
        version: components['schemas']['version'];
      }[];
      connections?: {
        name: string;
        version: components['schemas']['version'];
      }[];
      environment?: components['schemas']['environment'] & unknown;
      /** Format: date-time */
      readonly createdAt?: string;
      keyVersion?: components['schemas']['version'];
      /** @description OPA version used to generate the bundle */
      opaVersion?: string;
    };
    paginationResponse: {
      /**
       * Format: int32
       * @description total number of items
       */
      total: number;
    };
  };
// ...
};